test: remove TIFF decompression bomb reproduction test
Remove the decode_bomb_test.go reproduction test. The MaxDecodedResolution
cap test is preserved by moving it into decode_test.go.
